Separate ethernet::Inner into Config and State.
This is useful to allow the Config to be immutably borrowed from
many places at the same time State is mutably borrowed by the
packet processing code.

While not very useful on its own, this is one of many small refactors that
will make it possible to split the Ethernet code into Ethernet-specific
and IP-specific logic.